Modified Files:  	Makefile.am configure.in data/csboard.in  	resource/csboard.glade src/parser/PGNParser.cs  	src/viewer/ChessGameWidget.cs src/viewer/GameViewer.cs  	src/viewer/GamesListWidget.cs  Added Files:  	src/viewer/ChessGameInfoWidget.cs
Replaced the HTML widget with TreeView and Labels. Created a new file
ChessGameInfoWidget.cs for the class ChessGameInfoWidget which
will contain the details of the a chess game.
Also modified the layout so that the viewer will have three columns
one for games list, chess board and the game details each.